# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Documentação da API.",
        "title": "Retro Board API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/v1",
    "paths": {
...
        "/teams/:teamId/boards": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves all boards associated with a specific team.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Board"
                ],
                "summary": "List all boards for a team",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Team ID",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of boards",
                        "schema": {
                            "$ref": "#/definitions/board.ListAllBoardsEnvelop"
                        }
                    },
                    "400": {
                        "description": "Invalid request (e.g., missing parameters or validation error)",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Team not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorEnvelope"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a board associated with the specified team.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Board"
                ],
                "summary": "Create a new board",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Team ID",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Board creation data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateBoardRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Board successfully created"
                    },
                    "400": {
                        "description": "Invalid request (e.g., missing parameters or validation error)",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Team not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
        "/teams/:teamId/boards/:boardId/columns": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves all columns associated with a specific board.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Board"
                ],
                "summary": "List all columns for a board",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Team ID",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Board ID",
                        "name": "boardId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of columns",
                        "schema": {
                            "$ref": "#/definitions/board.ListAllColumnsEnvelop"
                        }
                    },
                    "400": {
                        "description": "Invalid request (e.g., missing parameters or invalid input data)",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorEnvelope"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a column associated with the specified board and team.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Board"
                ],
                "summary": "Create a new column",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Team ID",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Board ID",
                        "name": "boardId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Column creation data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateColumnRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Column successfully created"
                    },
                    "400": {
                        "description": "Invalid request (e.g., missing parameters or validation error)",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Team or board not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
        "/teams/:teamId/boards/:boardId/columns/:columnId/cards": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves all cards associated with a specific column on a board.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Board"
                ],
                "summary": "List all cards for a column on a board",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Team ID",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Board ID",
                        "name": "boardId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Column ID",
                        "name": "columnId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of cards",
                        "schema": {
                            "$ref": "#/definitions/board.ListAllCardsEnvelop"
                        }
                    },
                    "400": {
                        "description": "Invalid request (e.g., missing parameters or invalid input data)",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Board or column not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorEnvelope"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a card associated with the specified column, board, and team.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Board"
                ],
                "summary": "Create a new card",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Team ID",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Board ID",
                        "name": "boardId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Column ID",
                        "name": "columnId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Card creation data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateCardRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Card successfully created"
                    },
                    "400": {
                        "description": "Invalid request (e.g., missing parameters or validation error)",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Team, board, or column not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "board.ListAllBoardsEnvelop": {
            "type": "object",
            "properties": {
                "boards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.ListAllBoardsResponse"
                    }
                }
            }
        },
        "board.ListAllCardsEnvelop": {
            "type": "object",
            "properties": {
                "cards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.ListAllCardsResponse"
                    }
                }
            }
        },
        "board.ListAllColumnsEnvelop": {
            "type": "object",
            "properties": {
                "columns": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.ListAllColumnsResponse"
                    }
                }
            }
        },
...
        "dtos.CreateBoardRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "board1"
                }
            }
        },
        "dtos.CreateCardRequest": {
            "type": "object",
            "properties": {
                "text": {
                    "type": "string",
                    "example": "card1-text"
                }
            }
        },
        "dtos.CreateColumnRequest": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string",
                    "example": "#FFFFFF"
                },
                "name": {
                    "type": "string",
                    "example": "column1"
                }
            }
        },
...
        "dtos.ListAllBoardsResponse": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean",
                    "example": true
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "board1"
                },
                "team_id": {
                    "type": "integer",
                    "example": 2
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dtos.ListAllCardsResponse": {
            "type": "object",
            "properties": {
                "column_id": {
                    "type": "integer",
                    "example": 2
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "member_id": {
                    "type": "integer",
                    "example": 2
                },
                "text": {
                    "type": "string",
                    "example": "card1-text"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dtos.ListAllColumnsResponse": {
            "type": "object",
            "properties": {
                "board_id": {
                    "type": "integer",
                    "example": 2
                },
                "color": {
                    "type": "string",
                    "example": "#FFFFFF"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "column1"
                },
                "position": {
                    "type": "integer",
                    "example": 1
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "utils.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error message"
                }
            }
        }
    },
    "securityDefinitions": {
        "": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}